Refactor activity types
