+++ packages/ui/src/common/types/index.ts
@@ -1,5 +1,5 @@
 export * from './modalState'
 export * from './onTransactionDone'
-export { Activity } from './Activity'
+export * from './Activity'
 
 export type Address = string
